backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt==4.0.1
python-jose
python-dotenv
alembic
pydantic[email]
python-multipart
celery[redis]
docker
groq